pyproject.toml
[tool.poetry]
name = "auth-with-flask-and-jwt"
version = "0.1.0"
description = ""
authors = ["copyNdpaste <julsnutini@naver.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^2.7.0"
Flask-JWT-Extended = "^4.2.0"
Inject = "^4.3.1"
pydantic = "^1.8.1"
psycopg2-binary = "^2.8.6"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
black = "^21.5b1"
pytest = "^6.2.4"
pytest-xdist = "^2.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
